New commit:
getData.r now includes working codes which will both download the .pdf files from flutracking.net AND scrape the data out of those files.

The code for doing all the scraping is old but has been improved to make it simpler (i.e. including new functions to reduce the amount of code and simplifying the for loop).

Also, all of the regex expressions are now contained in the 'flutracking_regex_codes.RDS' file.